+++ src/tui/view/common/template_preview.rs
@@ -55,6 +55,13 @@ impl TemplatePreview {
             Self::Disabled { template }
         }
     }
+
+    pub fn template(&self) -> &Template {
+        match self {
+            Self::Disabled { template } => template,
+            Self::Enabled { template, .. } => template,
+        }
+    }
 }
 
 impl Generate for &TemplatePreview {